+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Configure default client and server packet filters that apply on startup and can be restored while tracing
+
+### Changed
+
+- Hide the trace client filter when only one client is present
+
 ## [1.8.2] - 2026-07-09
 
 ### Added